# Fil_Rouge-openapi_1.yaml
openapi: "3.1.0"
info:
  title: "Fil_Rouge API"
  description: "Fil_Rouge API"
  version: "1.0.0"
servers:
  - url: "https://Fil_Rouge"
paths:
...
  /api/retards/countRetardsByApprenant/{apprenantId}:
    get:
      summary: "GET api/retards/countRetardsByApprenant/{apprenantId}"
      operationId: "countRetardsByApprenant"
      parameters:
        - name: "apprenantId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "integer"
                format: "int64"
  /api/retards/deleteRetard/{id}:
    delete:
      summary: "DELETE api/retards/deleteRetard/{id}"
      operationId: "deleteRetard"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Void"
  /api/retards/getRetardById/{id}:
    get:
      summary: "GET api/retards/getRetardById/{id}"
      operationId: "getRetardById"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Retard"
  /api/retards/getRetardsByApprenant/{apprenantId}:
    get:
      summary: "GET api/retards/getRetardsByApprenant/{apprenantId}"
      operationId: "getRetardsByApprenant"
      parameters:
        - name: "apprenantId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Retard"
  /api/retards/saveRetard&app/{apprenantId}:
    post:
      summary: "POST api/retards/saveRetard&app/{apprenantId}"
      operationId: "saveRetard"
      parameters:
        - name: "apprenantId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Retard"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Retard"
  /api/retards/updateRetard/{id}:
    put:
      summary: "PUT api/retards/updateRetard/{id}"
      operationId: "updateRetard"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Retard"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Retard"
...
components:
  schemas:
    Apprenant:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        nom:
          type: "string"
        email:
          type: "string"
        password:
          type: "string"
        role:
          type: "string"
          enum:
            - "admin"
            - "apprenant"
...
    Void:
      type: "object"
      properties: { }
    Retard:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        date:
          type: "string"
          format: "date-time"
        durationDeRetard:
          type: "integer"
          format: "int32"
        etat_retard:
          type: "string"
          enum:
            - "JUSITIFIER"
            - "NONJUSITIFIER"
        apprenant:
          $ref: "#/components/schemas/Apprenant"
...